stylistic updates to forms.php
Added Ascii text indicators for which tables form data will
be placed into.  I am still trying to figure out how I will
utilize the two tables.  It should be noted that this is a
constraint of the assignment, and not a design I would choose.

The form itself may need to be restructured so that I can submit
the student info, and pull the auto-incremented user_id from the
student table into the stu_interest table.  user_id is the foreign key
in the stu_interest table.