creating plan array variables for db submit:
There is a lot of debugging code in this submit.
Most of it was trying to figure out how to find values
for day and night.  When I tackle the info array, I suspect
I will try a diffrent strategy.

(Most of this debugging will be removed as it serves no
higher purpose, and looks sloppy.)